Fix iOS Safari compatibility issue.
Safari on iOS was throwing "ReferenceError: Can't find variable: Notification"
on page load. By checking if Notification is undefined before referencing
it, the problem is avoided and the app loads without trouble.

My understanding from brief reading is that the firebase library
is not compatible with Safari on iOS, which is the root cause of
why Notification does not work on that browser.