Automatic Deck Sharing (#358)
* Fix downloading log when user cards are not available

* Remove Share deck button

Sharing deck is now automatic on each pick/card movement (main <-> sideboard)/basic lands change.

This generates a lot more traffic during the brewing phase, but I think it's managable.

It also generates many updates of the local draft log:
Currently local draft logs are compressed and stored in localStorage, since this is a pretty costly operation, I added a debouncing mechanism waiting 5sec. for the changes to stop before commiting to compressing and saving to disk.
This adds a complication as some changes may not be commited to disk when the user leaves the page. I added a warning when this occurs, but there's no perfect solution for this.

This also updates how DraftLogLive gets the decklists, it's no longer a special case but now uses the decklists embeded in the draft log, as it should have since the begining.

* Patch Note

* Typo

* Front tests: Skip the confirmation dialog when exiting/refreshing while there's a pending call to storeDraftLogs