Fix broken layout when reconnecting during a Team Sealed
Reconnecting with fixed deck enabled and disabling it caused a broken layout: a height style was
applied to the draft container. This is caused by a broken ref: For some reason vue sets the ref to
the parent the first time the style is applied (?!). I would have expected it to be undefined as long as
the DOM isn't completely mounted, I'm not sure what's going on.
Anyway, converting the condition from an 'if' to a 'show' avoids the situation,
probably because both elements are created at the same time now and the ref always points to the correct element.