Refactor booster generation error reporting.
Removed 'onError' property, generateBooster can now return a MessageError.
Some effort to propagate the error.
Final goal will be to remove usage of Session.emitError in favor of acknowledgment callbacks like in startTeamSealed (but this needs more work on the client side).

Also fixed a bug in startTeamSealed were session could be stuck in drafting state after a generateBooster error.