Update drag & drop between deck and sideboard for sortablejs-vue3
sortablejs-vue3 is a thinner wrapper around sortablejs than vuedragabble: Drag & Drop between distinct lists isn't well supported.
We have to handle updating the underlying list ourselves and delete the previous DOM element (sortablejs was intended to be used with DOM elements only) before vue renders the new one.

This patch does that for the deck, sideboard and collapsed sideboard.
This approach has some limits since it forces the parent of CardPool to know for sure where the card with the supplied UniqueCardID comes from.
Here it can only be this.deck or this.sideboard depending on the context, so this is not a problem, but it's not generic at all and it requires a weird gymnastic in CardPool to
allow the parent to update its `cards` prop before updating the column.
A more generic solution would be to attach the whole object to the DOM element (That's what vuedragabble does if I understood correctly), but I'm not sure how robust that is.
Might be worthy of a small library on top of sortablejs-vue3 if we need it elsewhere.