Make tsc emit definition files for the server and uses these in client side when possible.
Using the definition files instead of the original source file when possible is an easy way to
better decouple the client from the server. Building the client was pulling most of the server code for no reason
because of dependencies between the server files (we were importing a file just to get a type and ended up trying to
compile Cards.ts from the server for example). It would be even better to restructure the server code,
but it's a pain (I only tried, idk, 5 times today using different methods? :D Sharing code (not just types) between
client and server is desirable for some functionnality, and it's not easy to de really cleanly, at this point
I'm convinced the best is just to keep all shared code in the server and use relative includes client side, and maybe
setup an alias on the client, but aliases are messy and conventions differ between node, ts and webpack. Type
definitions works fine here because webpack can just ignore them so only the client-side ts needs to be able to find them).

There are still files that should obvisouly be restructured to avoid future mistakes.