Use FontAwesome Vue libraries instead of bundled fonts.
This will allow for a nice reduction of initial download size by selecting only the icons we actually need.
This patch should convert all icons to the new syntax, but I may have missed some...

There are still some instances of direct usage of the 'Font Awesome 5' font, so it's still included until I find a workaround for those (either a specialized font with only a few glyphs, or a markup update).