[Client] Bugfix: Keep sortable lists in sync with DOM.
When moving object within a list via drag & drop, sortable will not automatically update the
underlying array. This patch adds `update` events to actually keep the underlying lists in sync
with the DOM elements changes. This fixes lots of weird behaviors in CardPool.

There's also what seems to be a reactivity bug with SortableJS-vue3 (see
https://github.com/MaxLeiter/sortablejs-vue3/issues/57): The component was not always
correctly re-rendered when externally modified (for example, with multiple cards in the collapsed
sideboard, drag one card to the end of the sideboard then click on it to move it to the deck:
the card was correctly moved, but the SortableJS-vue3 component was not correctly updated resulting
in a ghost card staying at the bottom of the sideboard and leading to many errors.)
This patch also adds dynamic keys to all SortableJS-vue3 components to force re-renders when the length
of the list externally changes. This may not be enough in all cases (when swapping cards without changing
the length for example), but it's enough for now, and way simpler than computing a md5 or any
other unique identifier.