[Client] Note about a potential bug on Chrome
Hover state isn't always correctly cleared when hovering cards in Winston draft and moving the cursor to the bottom of the card.
The pointer-events: none; rule is correctly applied, but the card is still displayed with z-index:800.
We can't apply pointer-events: none to .card-column > .column-card-container > .card without disabling card zooming.
This looks like a Chrome bug as it works fine in Firefox, but idk.