MinesweeperDraft: Properly deserialize draft state when loading from disk.
MinesweeperCell and MinesweeperGrid instances were not properly restored when reloading from disk, corrupting the draft state.
This adds a deserialize method to MinesweeperDraftState that correctly instanciate MinesweeperCells and MinesweeperGrids as necessary.
This patch also adds tests for all users disconnecting and reconnecting during a Minesweeper Draft and simulating a server restart with full
draft state serialization (which should be extended to all game modes).