Removed exessive keying on CardPool columns.
This caused exessive complete re-rendering of columns, invalidating references to them.
This was particularly problematic when a force pick happened during a drag & drop,
causing a crash in SortableJS (see #606).

The issue isn't completely fixed, it can still happen specifically when dragging from
the column updated by the force pick (`oldIndex` becomes outdated), but it should be harder to trigger.

I really hope removing it won't bring back another cause of desync :))

As a side effect of the re-rendering, temporary style 'drop-active' was implicitly removed.
We now have to do it explicitly.