Bots: Use a local copy of booster in getScores.
The booster might be externally modified (e.g. by a player picking a card) while the bot request is in flight. This avoids reporting a false error in this case, and makes sure the fallback is called on the correct booster (although we probably don't care about the result anymore at this point).