Adjust mythic rate for each set
https://www.lethe.xyz/mtg/collation/collation-overview.html :
"If rares and mythic rares are printed on the same sheet, then each rare will be printed twice as many times on the sheet as each mythic rare."

So for sets where rares and mythics are printed on the same sheet, we can compute the expected mythic/rare ratio knowing how
many of each category are in the set. This lines up perfectly with the 1/7.4 ratio published by MaRo for ZNR, and with data from
MTGA for WoE (1/7).