Scryfall data update.
They removed the mb1 set (Mystery Booster) which we rely on. I don't have a clean solution for now, I added a backup of mb1 card data (we can't easily translate mb1 to plist because PaperBoosterFactory relies on mb1 and the card collector numbers, which all changed), this will surely break at some point, but we should be able to keep it around and update the card images if necessary.