Do not use the user current sessionID in removeUserFromSession
This should not matter, but previously if we end up in an inconsistent state where the
passed userID doesn't have a Connection (this seems to sometimes happen after connection issues),
or their sessionID doesn't match the passed session ID (never seen that one), but the user is
still somehow in the session, the session would be permanently bricked.
This change should allow the session owner to manually boot the disconnected player and hopefully
regain a consistent state, allowing the draft to continue with bots or even the disconnected player
to properly reconnect.